split status functions
