+++ modules/client/src/addresslistVoting/internal/utils.ts
@@ -2,9 +2,9 @@ import {
   decodeRatio,
   getCompactProposalId,
   hexToBytes,
+  InvalidProposalStatusError,
 } from "@aragon/sdk-common";
 import {
-  computeProposalStatus,
   SubgraphAction,
   SubgraphVoteValuesMap,
   VoteValues,
@@ -21,7 +21,11 @@ import {
   SubgraphAddresslistVotingProposalListItem,
   SubgraphAddresslistVotingVoterListItem,
 } from "./types";
-import { DaoAction, ProposalMetadata } from "@aragon/sdk-client-common";
+import {
+  DaoAction,
+  ProposalMetadata,
+  ProposalStatus,
+} from "@aragon/sdk-client-common";
 
 export function toAddresslistVotingProposal(
   proposal: SubgraphAddresslistVotingProposal,
@@ -143,4 +147,58 @@ export function addresslistVotingInitParamsToContract(
     votingSettingsToContract(params.votingSettings),
     params.addresses,
   ];
+}
+
+export function computeProposalStatus(
+  proposal:
+    | SubgraphAddresslistVotingProposal
+    | SubgraphAddresslistVotingProposalListItem,
+): ProposalStatus {
+  const now = new Date();
+  const startDate = new Date(
+    parseInt(proposal.startDate) * 1000,
+  );
+  const endDate = new Date(parseInt(proposal.endDate) * 1000);
+  if (proposal.executed) {
+    return ProposalStatus.EXECUTED;
+  }
+  if (startDate >= now) {
+    return ProposalStatus.PENDING;
+  }
+  if (proposal.potentiallyExecutable || proposal.earlyExecutable) {
+    return ProposalStatus.SUCCEEDED;
+  }
+  if (endDate >= now) {
+    return ProposalStatus.ACTIVE;
+  }
+  return ProposalStatus.DEFEATED;
+}
+
+export function computeProposalStatusFilter(status: ProposalStatus) {
+  let where = {};
+  const now = Math.round(new Date().getTime() / 1000).toString();
+  switch (status) {
+    case ProposalStatus.PENDING:
+      where = { startDate_gte: now };
+      break;
+    case ProposalStatus.ACTIVE:
+      where = { startDate_lt: now, endDate_gte: now, executed: false };
+      break;
+    case ProposalStatus.EXECUTED:
+      where = { executed: true };
+      break;
+    case ProposalStatus.SUCCEEDED:
+      where = { potentiallyExecutable: true, endDate_lt: now };
+      break;
+    case ProposalStatus.DEFEATED:
+      where = {
+        potentiallyExecutable: false,
+        endDate_lt: now,
+        executed: false,
+      };
+      break;
+    default:
+      throw new InvalidProposalStatusError();
+  }
+  return where;
 }
