+++ modules/client/src/tokenVoting/internal/utils.ts
@@ -1,5 +1,4 @@
 import {
-  computeProposalStatus,
   SubgraphAction,
   SubgraphVoteValuesMap,
   VoteValues,
@@ -34,10 +33,12 @@ import {
   decodeRatio,
   getCompactProposalId,
   hexToBytes,
+  InvalidProposalStatusError,
 } from "@aragon/sdk-common";
 import {
   DaoAction,
   ProposalMetadata,
+  ProposalStatus,
   TokenType,
 } from "@aragon/sdk-client-common";
 
@@ -275,4 +276,58 @@ export function toTokenVotingMember(
       };
     }),
   };
+}
+
+
+
+export function computeProposalStatus(
+  proposal: SubgraphTokenVotingProposal | SubgraphTokenVotingProposalListItem,
+): ProposalStatus {
+  const now = new Date();
+  const startDate = new Date(
+    parseInt(proposal.startDate) * 1000,
+  );
+  const endDate = new Date(parseInt(proposal.endDate) * 1000);
+  if (proposal.executed) {
+    return ProposalStatus.EXECUTED;
+  }
+  if (startDate >= now) {
+    return ProposalStatus.PENDING;
+  }
+  if (proposal.potentiallyExecutable || proposal.earlyExecutable) {
+    return ProposalStatus.SUCCEEDED;
+  }
+  if (endDate >= now) {
+    return ProposalStatus.ACTIVE;
+  }
+  return ProposalStatus.DEFEATED;
+}
+
+export function computeProposalStatusFilter(status: ProposalStatus) {
+  let where = {};
+  const now = Math.round(new Date().getTime() / 1000).toString();
+  switch (status) {
+    case ProposalStatus.PENDING:
+      where = { startDate_gte: now };
+      break;
+    case ProposalStatus.ACTIVE:
+      where = { startDate_lt: now, endDate_gte: now, executed: false };
+      break;
+    case ProposalStatus.EXECUTED:
+      where = { executed: true };
+      break;
+    case ProposalStatus.SUCCEEDED:
+      where = { potentiallyExecutable: true, endDate_lt: now };
+      break;
+    case ProposalStatus.DEFEATED:
+      where = {
+        potentiallyExecutable: false,
+        endDate_lt: now,
+        executed: false,
+      };
+      break;
+    default:
+      throw new InvalidProposalStatusError();
+  }
+  return where;
 }
